+++ src/test/ProtectedRoute.test.tsx
@@ -1,13 +1,13 @@
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
-import ProtectedRoute from './ProtectedRoute';
+import ProtectedRoute from '../components/ProtectedRoute';
 import { configureStore } from '@reduxjs/toolkit';
 import authReducer from '../features/auth/authSlice';
 import {
   CssVarsProvider,
 } from "@mui/joy";
-import { preloadedState as mockUserState } from '../../testMocks';
+import { preloadedState as mockUserState } from './testMocks';
 
   
   describe('Protected route', () => {
